requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dataclasses-json==0.5.14
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
Flask-RESTful==0.3.10
frozenlist==1.4.0
fsspec==2023.6.0
gunicorn==21.2.0
idna==3.4
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.266
langsmith==0.0.25
llama-index==0.8.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.1
Pygments==2.16.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
stack-data==0.6.2
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
yarl==1.9.2
